requirements.txt
acres==0.1.0
altair==5.5.0
attrs==24.2.0
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-speech==1.41.1
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
dateparser==1.2.0
etelemetry==0.3.1
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
httplib2==0.22.0
idna==3.10
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
narwhals==1.16.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==2.2.0
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pathvalidate==3.2.1
pillow==11.0.0
portalocker==2.10.1
protobuf==5.29.1
prov==2.0.1
puremagic==1.28
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
pydot==3.0.3
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
